add a command "set-keepalive" to turn on/off keepalive state
The "keepalive" state is a kernel feature for components and
devices that when a component or device is put into "keepalive"
state, the underlying hardware will be kept intact, no write is
permitted to happen on the hardware. Meanwhile, the software
instance that corresponds to the underlying hardware can be
destroyed and re-created. When re-created, the software instance
will be re-attached to the hardware state, and its private state
will be restored.

This patch, as preparation step, introduce a new Qemu command
which aims to check if the system support "keepalive" state or
not, and turn on/off the "keepalive" state of all relevant
components.

This command can be used for implementing "Qemu fast restart".

The goal of "Qemu fast restart" is to achieve quick Qemu exit,
upgrade, and restart while impose less impact to the VM. It can
be implemented by the following steps:

    1, Check if the system support "keepalive" state or not.
    2. Put all relevant components into "keepalive" state.
    3, Trigger keepalive-aware snapshot saving.
    4, Exit Qemu.
    5, Qemu can be upgraded.
    6, Relaunch Qemu in stopped mode.
    7, Trigger keepalive-aware snapshot loading.
    8, Validate the VM state to make sure no security issues.
    9, Clear "keepalive" state for all relevant components
    10, Resume VM running

The following is an example usage to implement "Qemu fast restart":

    # qemu-system-x86_64 --enable-kvm -M q35 -m 4G -smp 1 \
           -hda ubuntu-1904.qcow2 -monitor stdio \
           -object memory-backend-file,id=dimm0,size=4g,mem-path=/dev/dax0.0,share=on,pmem=on,align=2M \
           -numa node,memdev=dimm0,cpus=0 \
           -device vfio-pci,host=81:00.0,romfile= \
    QEMU 5.0.50 monitor - type 'help' for more information
    (qemu) migrate_set_capability x-ignore-shared on
    (qemu) stop
    (qemu) set-keepalive on,uuid=619cdf24-226f-4418-9f5a-98346019860e
    (qemu) savevm s0
    (qemu) q

    # qemu-system-x86_64 --enable-kvm -M q35 -m 4G -smp 1 \
           -hda ubuntu-1904.qcow2 -monitor stdio \
           -object memory-backend-file,id=dimm0,size=4g,mem-path=/dev/dax0.0,share=on,pmem=on,align=2M \
           -numa node,memdev=dimm0,cpus=0 \
           -device vfio-pci,host=81:00.0,keepalive_token=619cdf24-226f-4418-9f5a-98346019860e \
           -S
    QEMU 5.0.50 monitor - type 'help' for more information
    (qemu) migrate_set_capability x-ignore-shared on
    (qemu) loadvm s0
    (qemu) set-keepalive off
    (qemu) c
    (qemu)

Here "set-keepalive on/off" is the newly added command. It will put
all the passthrough devices and related components into keepalive
state.

Signed-off-by: Jason Zeng <jason.zeng@intel.com>